Round Robin 2.0

+++ Planificador/mainwindow.h
@@ -103,7 +103,7 @@ private:
     int entr;
     int esp;
 
-    double quantum;
+    int quantum;
 
     double R_promedio=0;
     double RN_promedio=0;
